Separate status and transcription as new view
